update error mapping

+++ src/sql/scripts/update_drift_alert.sql
@@ -2,4 +2,4 @@
 
 UPDATE scouter.drift_alerts
 SET status = $1
-WHERE id = $2;
+WHERE id = $2::integer;
